Share cats instances and test for Dual
