+++ core/src/test/scala/io/chrisdavenport/monoids/DualTests.scala
